avoid code repeatition
