+++ transformer_engine/pytorch/module/grouped_linear.py
@@ -2026,41 +2026,6 @@ class GroupedLinear(TransformerEngineBaseModule):
             weight_quantizers[i].internal = not self.primary_weights_in_fp8
         return weight_quantizers
 
-    def _configure_weight_quantizer_optimize_for_gemm(
-        self,
-        quantizer: Quantizer,
-        weight: torch.Tensor,
-    ) -> None:
-        """Configure preswizzling for the per-weight, single-tensor quantize kernel.
-
-        GroupedLinear uses ``group_quantize`` for activations, but its cached weights
-        are quantized individually by ``quantize_weight``. NVFP4 RHT eligibility
-        therefore uses the single-tensor kernel's 64-row alignment, not the grouped
-        kernel's 128-row alignment.
-        """
-        if self.primary_weights_in_fp8:
-            quantizer.optimize_for_gemm = False
-            return
-        if not isinstance(quantizer, NVFP4Quantizer):
-            quantizer.optimize_for_gemm = True
-            return
-
-        rows, cols = weight.numel() // weight.shape[-1], weight.shape[-1]
-        capability = get_device_compute_capability()
-        arch_supported = (10, 0) <= capability <= (11, 0)
-        if quantizer.with_rht:
-            enabled = arch_supported and rows % 64 == 0 and cols % 128 == 0
-        else:
-            enabled = (
-                arch_supported
-                and quantizer.with_2d_quantization
-                and not quantizer.row_scaled_nvfp4
-                and not quantizer.nvfp4_use_4over6
-                and rows % 128 == 0
-                and cols % 128 == 0
-            )
-        quantizer.optimize_for_gemm = enabled
-
     def _get_quantizers(self):
         weight_quantizers = self._get_weight_quantizers()
         input_quantizers, output_quantizers = (
